requirements.txt
# Pinned for Python 3.12 on Linux (CPU/GPU). Keep torch/torchaudio versions aligned.
torch==2.5.1
torchaudio==2.5.1

numpy==2.1.3
PyYAML==6.0.2
tqdm==4.67.1
rich==13.9.4
soundfile==0.12.1
jiwer==3.0.5
